Occupancy map v1.: Implement within InteractiveMarkerServer class
Motivation: InteractiveMarkerServer does the reference setting. So
we check if the marker is in occupied space before setting the ref.
If it is in occupied or unknown region, we don't set the reference
and we keep the last valid reference.
Shortcoming: Always keep the memory allocated for the map occupied.
Shortciming-2: In sheet-5 we won't use the marker anymore.So we
might want to implement it within Autopilot.cpp
Next version will do so. It will keep recieving references from the
marker even if it is invalid. Do the checking in ControllerCallback